작동해 주세요 Heroku!
Attribute Error는 이제 그만...
