+++ app.py
@@ -1,7 +1,6 @@
 from flask import Flask
 from flask_restful import Api
 
-import model
 import resource
 import config
 
@@ -18,6 +17,6 @@ api.add_resource(resource.BoardApi, '/board')
 api.add_resource(resource.PinListApi, '/pin/list')
 api.add_resource(resource.BoardListApi, '/board/list')
 
-if __name__ == '__main__':
-    model.initialize()
-    app.run(debug=config.DEBUG, host=config.HOST, port=config.PORT)
+#if __name__ == '__main__':
+#    model.initialize()
+#    app.run(debug=config.DEBUG, host=config.HOST, port=config.PORT)
