+++ apple.py
@@ -1,0 +1,9 @@
+from models.User import User
+from models import DB, Pin, Board
+
+
+# DB에 접속할 때 default Board를 자동으로 생성해주는 함수
+
+
+
+
